Add basic database functionality
I have set up the basics for the database needed to store user information. signup.html is an example sign-up form used to collect the user's login information: username, email, and password, and stores it in the MongoDB database. Currently, the database is local (i.e on Liam's machine), but we should be able to change this with some tweaks to the code.